+++ src/__tests__/useForm/setValue.test.tsx
@@ -553,6 +553,54 @@ describe('setValue', () => {
       expect(result.current.formState.errors?.test?.message).toBe('min');
     });
 
+    it('should validate input correctly with existing error', async () => {
+      const Component = () => {
+        const {
+          register,
+          setError,
+          setValue,
+          formState: { errors },
+        } = useForm();
+
+        return (
+          <>
+            <input {...register('test', { required: true })} />
+            <button
+              onClick={() => {
+                setError('test', { type: 'somethingWrong', message: 'test' });
+              }}
+            >
+              setError
+            </button>
+            <button
+              onClick={() => {
+                setValue('test', 'bill', {
+                  shouldValidate: true,
+                });
+              }}
+            >
+              update
+            </button>
+            <p>{errors?.test?.message}</p>
+          </>
+        );
+      };
+
+      render(<Component />);
+
+      fireEvent.click(screen.getByRole('button', { name: 'setError' }));
+
+      await waitFor(() => {
+        screen.getByText('test');
+      });
+
+      await actComponent(async () => {
+        await fireEvent.click(screen.getByRole('button', { name: 'update' }));
+      });
+
+      expect(screen.queryByText('test')).toBeNull();
+    });
+
     it('should not be called trigger method if options is empty', async () => {
       const { result } = renderHook(() => useForm<{ test: string }>());