+++ src/useForm.ts
@@ -363,7 +363,7 @@ export function useForm<
         )
       )[name];
 
-      shouldRenderBaseOnError(skipReRender, name, error);
+      await shouldRenderBaseOnError(skipReRender, name, error);
 
       return isUndefined(error);
     },
@@ -549,7 +549,7 @@ export function useForm<
     [criteriaMode, shouldUseNativeValidation],
   );
 
-  const setInternalValues = React.useCallback(
+  const setValues = React.useCallback(
     (
       name: FieldPath<TFieldValues>,
       value: UnpackNestedValue<
@@ -564,7 +564,7 @@ export function useForm<
 
         (isFieldArray || !isPrimitive(fieldValue) || (field && !field._f)) &&
         !isDateObject(fieldValue)
-          ? setInternalValues(
+          ? setValues(
               fieldName,
               fieldValue as SetFieldValue<TFieldValues>,
               options,
@@ -623,7 +623,7 @@ export function useForm<
     }
 
     ((field && !field._f) || isFieldArray) && !isNullOrUndefined(value)
-      ? setInternalValues(name, value, isFieldArray ? {} : options)
+      ? setValues(name, value, isFieldArray ? {} : options)
       : setFieldValue(name, value, options, true, !field);
 
     isFieldWatched(name) && subjectsRef.current.state.next({});
@@ -631,8 +631,7 @@ export function useForm<
   };
 
   const handleChange: ChangeHandler = React.useCallback(
-    async ({ type, target, target: { value, type: inputType } }) => {
-      let name = (target as Ref)!.name;
+    async ({ type, target, target: { value, name, type: inputType } }) => {
       let error;
       let isValid;
       const field = get(fieldsRef.current, name) as Field;
@@ -868,7 +867,9 @@ export function useForm<
           unset(formStateRef.current.dirtyFields, inputName);
         !options.keepTouched &&
           unset(formStateRef.current.touchedFields, inputName);
-        !options.keepDefaultValue && unset(defaultValuesRef.current, inputName);
+        !shouldUnregister &&
+          !options.keepDefaultValue &&
+          unset(defaultValuesRef.current, inputName);
       }
     }
 
@@ -1218,6 +1219,7 @@ export function useForm<
         formStateRef,
         defaultValuesRef,
         fieldArrayDefaultValuesRef,
+        setValues,
         unregister,
         shouldUnmount: shouldUnregister,
       }),
@@ -1232,7 +1234,7 @@ export function useForm<
     register,
     handleSubmit,
     watch: React.useCallback(watch, []),
-    setValue: React.useCallback(setValue, [setInternalValues]),
+    setValue: React.useCallback(setValue, [setValues]),
     getValues: React.useCallback(getValues, []),
     reset: React.useCallback(reset, []),
     clearErrors: React.useCallback(clearErrors, []),